+++ tiled/adapters/awkward_buffers.py
@@ -1,10 +1,8 @@
 """
 A directory containing awkward buffers, one file per form key.
 """
-import collections.abc
-from collections.abc import Buffer
 from pathlib import Path
-from typing import Any, Iterator, List, Optional, Union
+from typing import Any, List, Optional, Union
 
 import awkward.forms
 
@@ -13,82 +11,8 @@ from ..structures.awkward import AwkwardStructure
 from ..structures.core import Spec, StructureFamily
 from ..utils import path_from_uri
 from .awkward import AwkwardAdapter
+from .awkward_directory_container import DirectoryContainer
 from .type_alliases import JSON
-
-
-class DirectoryContainer(collections.abc.MutableMapping[str, Buffer]):
-    """ """
-
-    def __init__(self, directory: Path, form: Any):
-        """
-
-        Parameters
-        ----------
-        directory :
-        form :
-        """
-        self.directory = directory
-        self.form = form
-
-    def __getitem__(self, form_key: str) -> Buffer:
-        """
-
-        Parameters
-        ----------
-        form_key :
-
-        Returns
-        -------
-
-        """
-        with open(self.directory / form_key, "rb") as file:
-            return file.read()
-
-    def __setitem__(self, form_key: str, value: Buffer) -> None:
-        """
-
-        Parameters
-        ----------
-        form_key :
-        value :
-
-        Returns
-        -------
-
-        """
-        with open(self.directory / form_key, "wb") as file:
-            file.write(value)
-
-    def __delitem__(self, form_key: str) -> None:
-        """
-
-        Parameters
-        ----------
-        form_key :
-
-        Returns
-        -------
-
-        """
-        (self.directory / form_key).unlink(missing_ok=True)
-
-    def __iter__(self) -> Iterator[str]:
-        """
-
-        Returns
-        -------
-
-        """
-        yield from self.form.expected_from_buffers()
-
-    def __len__(self) -> int:
-        """
-
-        Returns
-        -------
-
-        """
-        return len(self.form.expected_from_buffers())
 
 
 class AwkwardBuffersAdapter(AwkwardAdapter):
